Move resource pool implementation part into impl
All namespace impl part moved into impl/resource_pool.h
